% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{as_density_distribution}
\alias{as_density_distribution}
\title{Takes a list of parameters and converts to a pdf}
\usage{
as_density_distribution(obj)
}
\arguments{
\item{obj}{the list to evaluate}
}
\value{
a function which takes in a vector x and returns P(x) for the defined distribution
}
\description{
Takes a list of parameters and converts to a pdf
}
