% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{as_evaled_expression}
\alias{as_evaled_expression}
\title{Evaluates an expression, returning a numeric value}
\usage{
as_evaled_expression(l)
}
\arguments{
\item{l}{the object (scalar or vector) to evaluate}
}
\value{
a float evaluation of the expression
}
\description{
Evaluates an expression, returning a numeric value
}
\examples{
as_evaled_expression(c("2+2", "9*9")) -> (4, 81)

}
