% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{as_random_distribution}
\alias{as_random_distribution}
\title{Evaluates an expression, returning a numeric value}
\usage{
as_random_distribution(obj)
}
\arguments{
\item{obj}{the string to evaluate}
}
\value{
a float evaluation of the expression
}
\description{
Evaluates an expression, returning a numeric value
}
\examples{
as_evaled_expression("2+2") -> 4

}
