% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{$.config}
\alias{$.config}
\title{Overrides the $ operator for S3 'config' objects to ensure that named args exist.}
\usage{
\method{$}{config}(x, name)
}
\description{
Overrides the $ operator for S3 'config' objects to ensure that named args exist.
}
