% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{data}
\name{config}
\alias{config}
\title{Map of configuration values, constructed via a call to `load_config`}
\format{
An object of class \code{logical} of length 1.
}
\usage{
config
}
\description{
Map of configuration values, constructed via a call to `load_config`
}
\keyword{datasets}
