% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_paths.R
\name{create_prefix}
\alias{create_prefix}
\title{create_prefix}
\usage{
create_prefix(..., prefix = "", sep = "-", trailing_separator = "")
}
\arguments{
\item{sep}{A character to use to separate different components of the scenario in the tag.  This argument cannot appear in any of the  .dots arguments.}

\item{.dots}{A set of strings, or lists of value,format (see sprintf).}
}
\description{
Function for creating scenario tags from components; Intended for use in filename construction
}
