% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUtils.R
\name{download_CSSE_global_data}
\alias{download_CSSE_global_data}
\title{Download CSSE global data}
\usage{
download_CSSE_global_data(filename, url, value_col_name)
}
\arguments{
\item{filename}{where case data will be stored}

\item{url}{URL to CSV on CSSE website}
}
\value{
data frame
}
\description{
Downloads the CSSE global case and death count data
}
