% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUtils.R
\name{download_CSSE_US_data}
\alias{download_CSSE_US_data}
\title{Download CSSE US data}
\usage{
download_CSSE_US_data(filename, url, value_col_name, incl_unassigned = FALSE)
}
\arguments{
\item{filename}{where case data will be stored}

\item{url}{URL to CSV on CSSE website}

\item{value_col_name}{Confirmed or Deaths}

\item{incl_unassigned}{Includes data unassigned to county-level (default is FALSE)}
}
\value{
data frame
}
\description{
Downloads the CSSE US case and death count data
}
