% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUtils.R
\name{download_reichlab_data}
\alias{download_reichlab_data}
\title{Download Reich Lab data}
\usage{
download_reichlab_data(filename, url, value_col_name)
}
\arguments{
\item{filename}{where case data will be stored}

\item{url}{URL to CSV on Reich Lab website}
}
\value{
data frame
}
\description{
Downloads the Reich Lab's US case and death count data
}
