% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUtils.R
\name{download_USAFacts_data}
\alias{download_USAFacts_data}
\title{Download USAFacts data}
\usage{
download_USAFacts_data(filename, url, value_col_name, incl_unassigned = FALSE)
}
\arguments{
\item{filename}{where case data will be stored}

\item{url}{URL to CSV on USAFacts website}

\item{value_col_name}{Confirmed or Deaths}

\item{incl_unassigned}{Includes data unassigned to counties (default is FALSE)}
}
\value{
data frame
}
\description{
Downloads the USAFacts case and death count data
}
