% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUtils.R
\name{get_CSSE_global_data}
\alias{get_CSSE_global_data}
\title{Pull global case and death count data from JHU CSSE}
\usage{
get_CSSE_global_data(
  case_data_filename = "data/case_data/jhucsse_case_data_crude.csv",
  death_data_filename = "data/case_data/jhucsse_death_data_crude.csv",
  append_wiki = TRUE
)
}
\value{
the case and deaths data frame
}
\description{
Pulls the CSSE cumulative case count and death data. Calculates incident counts.
}
\details{
Data preview:
 $ Update     : Date "2019-12-01" "2019-12-02" ...
 $ UID        : chr "AFG" "AFG" "AFG" "AFG" ...
 $ iso2       : chr  "AF" "AF" "AF" "AF" ...
 $ iso3       : chr  "AFG" "AFG" "AFG" "AFG" ...
 $ Latitude   : num  "AF" "AF" "AF" "AF" ...
 $ Longitude  : num  "AFG" "AFG" "AFG" "AFG" ...
 $ Confirmed  : num  0 0 0 0 0 0 0 0 0 0 ...
 $ Deaths     : num  0 0 0 0 0 0 0 0 0 0 ...
 $ incidI     : num  0 0 0 0 0 0 0 0 0 0 ...
 $ incidDeath : num  0 0 0 0 0 0 0 0 0 0 ...
 $ Country_Region  : chr  "Afghanistan" "Afghanistan" "Afghanistan" "Afghanistan" ...
 $ Province_State  : chr NA NA NA NA ...
 $ source     : chr "AFG" "AFG" "AFG" "AFG" ...
}
