% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUtils.R
\name{get_CSSE_US_data}
\alias{get_CSSE_US_data}
\title{Pull US case and death count data from JHU CSSE}
\usage{
get_CSSE_US_data(
  case_data_filename = "data/case_data/jhucsse_us_case_data_crude.csv",
  death_data_filename = "data/case_data/jhucsse_us_death_data_crude.csv",
  incl_unassigned = FALSE
)
}
\arguments{
\item{case_data_filename}{where case data will be stored}

\item{death_data_filename}{where death data will be stored}

\item{incl_unassigned}{Includes data unassigned to county-level (default is FALSE)}
}
\value{
the case and deaths data frame
}
\description{
Pulls the CSSE cumulative case count and death data. Calculates incident counts.
}
\details{
Returned data preview:
tibble [352,466 × 7] (S3: grouped_df/tbl_df/tbl/data.frame)
 $ FIPS       : chr [1:352466] "00001" "00001" "00001" "00001" ...
 $ source     : chr [1:352466] "NY" "NY" "NY" "NY" ...
 $ Update     : Date[1:352466], format: "2020-01-22" "2020-01-23" ...
 $ Confirmed  : num [1:352466] 0 0 0 0 0 0 0 0 0 0 ...
 $ Deaths     : num [1:352466] 0 0 0 0 0 0 0 0 0 0 ...
 $ incidI     : num [1:352466] 0 0 0 0 0 0 0 0 0 0 ...
 $ incidDeath : num [1:352466] 0 0 0 0 0 0 0 0 0 0 ...
}
