% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUtils.R
\name{get_CSSE_US_matchGlobal_data}
\alias{get_CSSE_US_matchGlobal_data}
\title{Pull CSSE US data in format similar to that of global data}
\usage{
get_CSSE_US_matchGlobal_data()
}
\value{
data frame
}
\description{
Pulls the CSSE US confirmed cases and deaths and calculates incident cases and deaths, putting them into a verbose geographic format like that with the global dataframe
}
