% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUtils.R
\name{get_groundtruth_from_source}
\alias{get_groundtruth_from_source}
\title{Wrapper function to pull data from different sources}
\usage{
get_groundtruth_from_source(
  source = "csse",
  scale = "US county",
  variables = c("Confirmed", "Deaths", "incidI", "incidDeath"),
  incl_unass = FALSE,
  get_hosp = FALSE
)
}
\arguments{
\item{source}{name of data source: reichlab, usafacts, csse}

\item{scale}{geographic scale: US county, US state, country (csse only), complete (csse only)}

\item{variables}{vector that may include one or more of the following variable names: Confirmed, Deaths, incidI, incidDeath, (hhsCMU source only: incidH_confirmed, incidH_all, hospCurr_confirmed, hospCurr_all)}
}
\value{
data frame
}
\description{
Pulls a groundtruth dataset with the variables specified
}
