% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUtils.R
\name{get_hhsCMU_allHosp_st_data}
\alias{get_hhsCMU_allHosp_st_data}
\title{Pull all state-level hospitalization data from HHS through CMU Delphi API}
\usage{
get_hhsCMU_allHosp_st_data(
  startdate = "20200101",
  enddate = stringr::str_remove_all(as.character(Sys.Date()), "-")
)
}
\value{
the full hospitalization data frame
}
\description{
Returned data preview:
Variables: 56
$ state                                         <chr> "AL", "AL", "AL",...
$ issue                                         <int> 20201129, 2020112...
$ date                                          <int> 20201001, 2020100...
$ hospital_onset_covid                          <int> 5, 8, 7, 4, 7, 11...
$ hospital_onset_covid_coverage                 <int> 110, 111, 110, 11...
$ inpatient_beds                                <int> 14925, 14590, 151...
$ inpatient_beds_coverage                       <int> 111, 109, 110, 11...
$ inpatient_beds_used                           <int> 11022, 10651, 104...
$ inpatient_beds_used_coverage                  <int> 111, 109, 110, 11...
$ inpatient_beds_used_covid                               <int> 962, 939, 931, 96...
$ inpatient_beds_used_covid_coverage                      <int> 110, 111, 110, 11...
$ previous_day_admission_adult_covid_confirmed            <int> 119, 105, 114, 95...
$ previous_day_admission_adult_covid_confirmed_coverage   <int> 112, 113, 112, 11...
$ previous_day_admission_adult_covid_suspected            <int> 102, 120, 108, 84...
$ previous_day_admission_adult_covid_suspected_coverage   <int> 110, 111, 110, 11...
$ previous_day_admission_pediatric_covid_confirmed        <int> 2, 1, 6, 3, 4, 2,...
$ previous_day_admission_pediatric_covid_confirmed_coverage        <int> 110, 111...
$ previous_day_admission_pediatric_covid_suspected        <int> 3, 9, 5, 10, 5, 1...
$ previous_day_admission_pediatric_covid_suspected_coverage        <int> 110, 111...
$ staffed_adult_icu_bed_occupancy                         <int> 1219, 1234, 1225,...
$ staffed_adult_icu_bed_occupancy_coverage                <int> 111, 109, 110, 11...
$ staffed_icu_adult_patients_confirmed_suspected_covid    <int> 289, 295, 307, 31...
$ staffed_icu_adult_patients_confirmed_suspected_covid_coverage    <int> 110, 111...
$ staffed_icu_adult_patients_confirmed_covid              <int> 275, 281, 287, 29...
$ staffed_icu_adult_patients_confirmed_covid_coverage     <int> 112, 113, 112, 11...
$ total_adult_patients_hosp_confirmed_suspected_covid     <int> 949, 925, 912, 94...
$ total_adult_patients_hosp_confirmed_suspected_covid_coverage     <int> 110, 111...
$ total_adult_patients_hosp_confirmed_covid               <int> 785, 777, 759, 77...
$ total_adult_patients_hosp_confirmed_covid_coverage      <int> 112, 113, 112, 11...
$ total_pediatric_patients_hosp_confirmed_suspected_covid <int> 13, 14, 19, 18, 1...
$ total_pediatric_patients_hosp_confirmed_suspected_covid_coverage <int> 110, 111...
$ total_pediatric_patients_hosp_confirmed_covid           <int> 7, 5, 11, 10, 10,...
$ total_pediatric_patients_hosp_confirmed_covid_coverage  <int> 110, 111, 110, 11...
$ total_staffed_adult_icu_beds                  <int> 1476, 1494, 1522,...
$ total_staffed_adult_icu_beds_coverage         <int> 111, 109, 110, 11...
$ inpatient_beds_utilization_coverage           <int> 111, 109, 110, 11...
$ inpatient_beds_utilization_numerator          <int> 11022, 10651, 104...
$ inpatient_beds_utilization_denominator        <int> 14925, 14590, 151...
$ percent_of_inpatients_with_covid_coverage     <int> 108, 107, 107, 10...
$ percent_of_inpatients_with_covid_numerator    <int> 937, 892, 927, 95...
$ percent_of_inpatients_with_covid_denominator  <int> 10908, 10558, 103...
$ inpatient_bed_covid_utilization_coverage      <int> 108, 107, 107, 10...
$ inpatient_bed_covid_utilization_numerator     <int> 937, 892, 927, 95...
$ inpatient_bed_covid_utilization_denominator   <int> 14735, 14431, 148...
$ adult_icu_bed_covid_utilization_coverage      <int> 108, 107, 107, 10...
$ adult_icu_bed_covid_utilization_numerator     <int> 280, 285, 307, 31...
$ adult_icu_bed_covid_utilization_denominator   <int> 1438, 1456, 1474,...
$ adult_icu_bed_utilization_coverage            <int> 111, 109, 110, 11...
$ adult_icu_bed_utilization_numerator           <int> 1219, 1234, 1225,...
$ adult_icu_bed_utilization_denominator         <int> 1476, 1494, 1522,...
$ inpatient_beds_utilization                    <dbl> 0.7384925, 0.7300...
$ percent_of_inpatients_with_covid              <dbl> 0.08590026, 0.084...
$ inpatient_bed_covid_utilization               <dbl> 0.06359009, 0.061...
$ adult_icu_bed_covid_utilization               <dbl> 0.1947149, 0.1957...
$ adult_icu_bed_utilization                     <dbl> 0.8258808, 0.8259...
$ Update                                        <date> 2020-10-01, 2020...
}
