% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUtils.R
\name{get_hhsCMU_hospCurr_st_data}
\alias{get_hhsCMU_hospCurr_st_data}
\title{Clean and trim state-level hospCurr data from HHS through CMU Delphi API}
\usage{
get_hhsCMU_hospCurr_st_data(
  startdate = "20200101",
  enddate = stringr::str_remove_all(as.character(Sys.Date()), "-")
)
}
\description{
Returned data preview: Update, FIPS, source, hospCurr_confirmed, hospCurr_all
}
