% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUtils.R
\name{get_hhsCMU_incidH_st_data}
\alias{get_hhsCMU_incidH_st_data}
\title{Clean and trim state-level incidH data from HHS through CMU Delphi API
Update is shifted one day back because data was originally reported as previous day hospital admissions
Returned data preview: Update, FIPS, source, incidH_confirmed, incidH_all}
\usage{
get_hhsCMU_incidH_st_data(
  startdate = "20200101",
  enddate = stringr::str_remove_all(as.character(Sys.Date()), "-")
)
}
\description{
Clean and trim state-level incidH data from HHS through CMU Delphi API
Update is shifted one day back because data was originally reported as previous day hospital admissions
Returned data preview: Update, FIPS, source, incidH_confirmed, incidH_all
}
