% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUtils.R
\name{get_islandareas_data}
\alias{get_islandareas_data}
\title{Pulls island area data from NY Times}
\usage{
get_islandareas_data()
}
\value{
the case data frame
}
\description{
Downloads data from NY Times and filters on island area
}
\details{
Returned data preview:
tibble [198 × 5] (S3: spec_tbl_df/tbl_df/tbl/data.frame)
 $ Update   : Date[1:198], format: "2020-03-13" "2020-03-14" "2020-03-14" ...
 $ source   : chr [1:198] "PR" "PR" "VI" "GU" ...
 $ FIPS     : chr [1:198] "72000" "72000" "78000" "66000" ...
 $ Confirmed: num [1:198] 3 4 1 3 5 1 3 5 2 3 ...
 $ Deaths   : num [1:198] 0 0 0 0 0 0 0 0 0 0 ...
}
