% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUtils.R
\name{get_reichlab_cty_data}
\alias{get_reichlab_cty_data}
\title{Pull county-level case and death count data from Reich Lab}
\usage{
get_reichlab_cty_data(
  cum_case_filename = "data/case_data/rlab_cum_case_data.csv",
  cum_death_filename = "data/case_data/rlab_cum_death_data.csv",
  inc_case_filename = "data/case_data/rlab_inc_case_data.csv",
  inc_death_filename = "data/case_data/rlab_inc_death_data.csv"
)
}
\value{
the case and deaths data frame
}
\description{
Pulls the Reich Lab incident and cumulative case count and death data.
}
\details{
Returned data preview:
tibble
 $ Update     : Date "2020-01-22" "2020-01-23" ...
 $ Confirmed  : num 0 0 0 0 0 0 0 0 0 0 ...
 $ Deaths     : num 0 0 0 0 0 0 0 0 0 0 ...
 $ incidI     : num 0 0 0 0 0 0 0 0 0 0 ...
 $ incidDeath : num 0 0 0 0 0 0 0 0 0 0 ...
 $ FIPS       : chr "01001" "01001" "01001" ...
 $ source     : chr "NY" "NY" "NY" "NY" ...
}
