% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUtils.R
\name{get_USAFacts_data}
\alias{get_USAFacts_data}
\title{Pull case and death count data from USAFacts}
\usage{
get_USAFacts_data(
  case_data_filename = "data/case_data/USAFacts_case_data.csv",
  death_data_filename = "data/case_data/USAFacts_death_data.csv",
  incl_unassigned = FALSE
)
}
\arguments{
\item{case_data_filename}{Filename where case data are stored}

\item{death_data_filename}{Filename where death data are stored}

\item{incl_unassigned}{Includes data unassigned to counties (default is FALSE)}
}
\value{
the case and deaths data frame
}
\description{
Pulls the USAFacts cumulative case count and death data. Calculates incident counts.
USAFacts does not include data for all the territories (aka island areas). These data are pulled from NYTimes.
}
\details{
Returned data preview:
tibble [352,466 × 7] (S3: grouped_df/tbl_df/tbl/data.frame)
 $ FIPS       : chr [1:352466] "00001" "00001" "00001" "00001" ...
 $ source     : chr [1:352466] "NY" "NY" "NY" "NY" ...
 $ Update     : Date[1:352466], format: "2020-01-22" "2020-01-23" ...
 $ Confirmed  : num [1:352466] 0 0 0 0 0 0 0 0 0 0 ...
 $ Deaths     : num [1:352466] 0 0 0 0 0 0 0 0 0 0 ...
 $ incidI     : num [1:352466] 0 0 0 0 0 0 0 0 0 0 ...
 $ incidDeath : num [1:352466] 0 0 0 0 0 0 0 0 0 0 ...
}
