% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_paths.R
\name{hpar_file_path}
\alias{hpar_file_path}
\title{Function for determining where to write the hospitalization parameter files}
\usage{
hpar_file_path(config, index, scenario, deathrate, create_directory = TRUE)
}
\arguments{
\item{config}{The config for this run}

\item{index}{The index of this simulation}

\item{scenario}{The scenario of the simulation}

\item{deathrate}{Name of deathrate scenario}
}
\description{
Function for determining where to write the hospitalization parameter files
}
