% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{load_config}
\alias{load_config}
\title{Returns a map of configuration loaded from the config YAML}
\usage{
load_config(fname)
}
\arguments{
\item{fname}{Load configuration from fname (optional, otherwise loads from CONFIG_PATH env var)}
}
\description{
Returns a map of configuration loaded from the config YAML
}
\examples{
config$parameters_seir$gamma

}
