% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_paths.R
\name{seeding_file_path}
\alias{seeding_file_path}
\title{Function for determining where to write the seeding.csv file}
\usage{
seeding_file_path(config, index, create_directory = TRUE)
}
\arguments{
\item{config}{The config for this run}

\item{index}{The index of this simulation}
}
\description{
Function for determining where to write the seeding.csv file
}
