% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_paths.R
\name{simulation_file_path}
\alias{simulation_file_path}
\title{Function for determining where to write the SEIR output to file}
\usage{
simulation_file_path(config, index, scenario, create_directory = TRUE)
}
\arguments{
\item{config}{The config for this run}

\item{index}{The index of this simulation}

\item{scenario}{The scenario of the simulation}
}
\description{
Function for determining where to write the SEIR output to file
}
