% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_paths.R
\name{snpi_file_path}
\alias{snpi_file_path}
\title{Function for determining where to write the seeding.csv file}
\usage{
snpi_file_path(config, index, scenario, create_directory = TRUE)
}
\arguments{
\item{config}{The config for this run}

\item{index}{The index of this simulation}

\item{scenario}{The scenario of the simulation}
}
\description{
Function for determining where to write the seeding.csv file
}
